refactor(modules): Move __main__.py content to cli.py
